+++ src/app/features/poll/components/vote/vote.component.spec.ts
@@ -1,25 +1,24 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { combineReducers, StoreModule } from '@ngrx/store';
+import { combineReducers, Store, StoreModule } from '@ngrx/store';
 
-import { OptionComponent } from '../option/option.component';
 import { VoteComponent } from './vote.component';
+
+import * as voteActions from '../../actions/vote.actions';
 
 import * as fromRoot from '../../../../reducers/root.reducer';
 import * as fromPoll from '../../reducers/root.reducer';
 
-import { poll } from '../../../../test-helpers';
+import { poll, user } from '../../../../test-helpers';
 
 describe('VoteComponent', () => {
   let component: VoteComponent;
   let fixture: ComponentFixture<VoteComponent>;
+  let store: Store<fromRoot.State>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [
-        VoteComponent,
-        OptionComponent
-      ],
+      declarations: [VoteComponent],
       imports: [
         StoreModule.forRoot({
           ...fromRoot.reducers,
@@ -34,6 +33,10 @@ describe('VoteComponent', () => {
     component = fixture.componentInstance;
     component.poll = poll.testPoll;
     component.pollId = 'Some PollId';
+
+    store = TestBed.get(Store);
+    spyOn(store, 'dispatch').and.callThrough();
+
     fixture.detectChanges();
   });
 
@@ -41,10 +44,39 @@ describe('VoteComponent', () => {
     expect(component).toBeTruthy();
   });
 
+  it('Should disable vote button', () => {
+    const voteButton = fixture.nativeElement.querySelector('.vote button');
+    expect(voteButton.disabled).toBe(true);
+  });
+
   describe('getPollOptions', () => {
     it('Should return the list of properties', () => {
       const pollOptions = component.getPollOptions();
       expect(pollOptions).toEqual(['Chipotle', 'Sheetz', 'Pulp']);
+    });
+  });
+
+  describe('When option is selected', () => {
+    it('Should dispatch SetVoteOption', () => {
+      const firstOption = fixture.nativeElement.querySelector('.vote .options li:first-child input');
+      firstOption.dispatchEvent(new Event('change'));
+      expect(store.dispatch).toHaveBeenCalledWith(new voteActions.SetVoteOption('Chipotle'));
+    });
+  });
+
+  describe('When voteButton is clicked', () => {
+    it('Should dispatch Vote', () => {
+      component.selectedOption = 'Chipotle';
+      fixture.detectChanges();
+
+      const voteButton = fixture.nativeElement.querySelector('.vote button');
+      voteButton.click();
+
+      expect(store.dispatch).toHaveBeenCalledWith(new voteActions.Vote({
+        pollId: 'Some PollId',
+        option: 'Chipotle',
+        userId: user.testUser.uid
+      }));
     });
   });
 });
