+++ src/app/features/poll/effects/vote.effects.ts
@@ -6,7 +6,6 @@ import { of } from 'rxjs';
 import { catchError, map, switchMap } from 'rxjs/operators';
 
 import { PollService } from '../../../shared/services/poll.service';
-import { VoteService } from '../services/vote.service';
 
 import * as appActions from '../../../actions/app.actions';
 import * as voteInfoActions from '../actions/vote-info.actions';
@@ -15,7 +14,7 @@ import * as voteActions from '../actions/vote.actions';
 @Injectable()
 export class VoteEffects {
 
-  constructor(private actions$: Actions, private pollService: PollService, private voteService: VoteService) { }
+  constructor(private actions$: Actions, private pollService: PollService) { }
 
   @Effect() vote$ =
     this.actions$
@@ -34,26 +33,4 @@ export class VoteEffects {
       .pipe(
         map(action => action as voteActions.VoteSucceeded),
         map(action => new voteInfoActions.TrackVote(action.payload)));
-
-
-  @Effect() trackVote$ =
-    this.actions$
-      .ofType(voteInfoActions.TRACK_VOTE)
-      .pipe(
-        map(action => action as voteInfoActions.TrackVote),
-        map(action => action.payload),
-        switchMap(payload => this.voteService.trackVote(payload)
-          .pipe(
-            map(data => new voteInfoActions.TrackVoteSucceeded(data)),
-            catchError(err => of(new appActions.Error(voteInfoActions.TRACK_VOTE, err.message))))));
-
-  @Effect() loadVoteInfo$ =
-    this.actions$
-      .ofType(voteInfoActions.LOAD_VOTE_INFO)
-      .pipe(
-        map(action => action as voteInfoActions.LoadVoteInfo),
-        switchMap(action => this.voteService.getVotesForUser(action.userId)
-          .pipe(
-            map(votes => new voteInfoActions.LoadVoteInfoSucceeded(votes)),
-            catchError(err => of(new appActions.Error(voteInfoActions.LOAD_VOTE_INFO, err.message))))));
 }
