+++ src/app/features/poll/poll.component.spec.ts
@@ -5,7 +5,6 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { combineReducers, Store, StoreModule } from '@ngrx/store';
 
 import { AlreadyVotedComponent } from './components/already-voted/already-voted.component';
-import { OptionComponent } from './components/option/option.component';
 import { VoteComponent } from './components/vote/vote.component';
 import { PollComponent } from './poll.component';
 
@@ -27,7 +26,6 @@ describe('PollComponent', () => {
       declarations: [
         PollComponent,
         VoteComponent,
-        OptionComponent,
         AlreadyVotedComponent
       ],
       imports: [
